backend/workers/requirements-gpu.txt
# Core AI libraries
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0
xformers>=0.0.22
safetensors>=0.4.0
bitsandbytes>=0.41.0

# LoRA training
peft>=0.6.0
kohya-ss>=0.0.1
datasets>=2.14.0

# ComfyUI dependencies
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0
matplotlib>=3.7.0
tqdm>=4.66.0
psutil>=5.9.0
requests>=2.31.0
aiohttp>=3.8.0
websockets>=11.0.0
pyyaml>=6.0.0
omegaconf>=2.3.0
einops>=0.7.0
kornia>=0.7.0
controlnet-aux>=0.0.6
insightface>=0.7.3
onnx>=1.14.0
onnxruntime>=1.16.0
opencv-contrib-python>=4.8.0
segment-anything>=1.0
ultralytics>=8.0.0
timm>=0.9.0
lpips>=0.1.4
pytorch-lightning>=2.0.0
rich>=13.0.0
pygit2>=1.12.0
tensorboard>=2.14.0
protobuf>=4.24.0

# AnimateDiff
animatediff>=0.0.1
av>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.0

# Worker dependencies
redis>=5.0.0
boto3>=1.34.0
structlog>=23.0.0
pydantic>=2.5.0
httpx>=0.25.0

# Hugging Face
huggingface-hub>=0.19.0
tokenizers>=0.14.0

# Additional utilities
wandb>=0.16.0
tensorboard>=2.14.0
tqdm>=4.66.0
click>=8.1.0
typer>=0.9.0


